[Task #1652] Remove TermAssignment aggregated into DTO.
Move term occurrence aggregated info retrieval into TermOccurrenceDao.
