+++ src/main/java/cz/cvut/kbss/termit/dto/assignment/TermOccurrences.java
@@ -1,27 +1,25 @@
 /**
  * TermIt
  * Copyright (C) 2019 Czech Technical University in Prague
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- *
+ * <p>
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 package cz.cvut.kbss.termit.dto.assignment;
 
 
-import cz.cvut.kbss.jopa.model.annotations.ConstructorResult;
-import cz.cvut.kbss.jopa.model.annotations.OWLDataProperty;
-import cz.cvut.kbss.jopa.model.annotations.SparqlResultSetMapping;
-import cz.cvut.kbss.jopa.model.annotations.VariableResult;
+import cz.cvut.kbss.jopa.model.annotations.*;
+import cz.cvut.kbss.jopa.vocabulary.RDFS;
 import cz.cvut.kbss.termit.util.Vocabulary;
 
 import java.net.URI;
@@ -43,9 +41,12 @@ import java.util.Objects;
                 @VariableResult(name = "suggested", type = Boolean.class)
         }
 ))
-public class TermOccurrences extends TermAssignments {
+public class TermOccurrences extends AbstractAssignmentsInfo {
 
     public static final String COUNT_PROPERTY = "http://onto.fel.cvut.cz/ontologies/application/termit/pojem/počet";
+
+    @OWLAnnotationProperty(iri = RDFS.LABEL)
+    private String resourceLabel;
 
     @OWLDataProperty(iri = COUNT_PROPERTY)
     private Integer count;
@@ -53,14 +54,24 @@ public class TermOccurrences extends TermAssignments {
     public TermOccurrences() {
     }
 
-    public TermOccurrences(URI term, URI resource, String resourceLabel, Integer count, String type, Boolean suggested) {
-        super(term, resource, resourceLabel, suggested);
+    public TermOccurrences(URI term, URI resource, String resourceLabel, Integer count, String type,
+                           Boolean suggested) {
+        super(term, resource);
+        this.resourceLabel = resourceLabel;
         this.count = count;
         addType(type);
         addType(Vocabulary.s_c_vyskyt_termu);
         if (suggested) {
             addType(Vocabulary.s_c_navrzeny_vyskyt_termu);
         }
+    }
+
+    public String getResourceLabel() {
+        return resourceLabel;
+    }
+
+    public void setResourceLabel(String resourceLabel) {
+        this.resourceLabel = resourceLabel;
     }
 
     public Integer getCount() {
@@ -83,19 +94,20 @@ public class TermOccurrences extends TermAssignments {
             return false;
         }
         TermOccurrences that = (TermOccurrences) o;
-        return Objects.equals(count, that.count);
+        return Objects.equals(resourceLabel, that.resourceLabel) && Objects.equals(count, that.count);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), count);
+        return Objects.hash(super.hashCode(), resourceLabel, count);
     }
 
     @Override
     public String toString() {
         return "TermOccurrences{" +
                 super.toString() +
-                "count=" + count +
+                ", resourceLabel='" + resourceLabel +
+                "', count=" + count +
                 "}";
     }
 }
